BIG UPDATE! Added office zone and changed quiz attribbutes.
3 cameras divide up the screen. With quiz area, office area, and cyber area.
Office has been designed and character in office space moves around via left click in that zone. Workers have a ping that can be right clicked on to pull up a random question from the quiz.
Start button has been changed to move toggle the camera.
Some of the questions have been changed so that they restart the game when clicked. Added player prefs to save values in the player prefs. This is so when the scene is reloaded everything stays the same.